ci(security): introduce secret scanning (gitleaks) + local workflow test + deep secret heuristics (#126)
* docs(readme): remove broken links and fix build/test section\n\n[AIR-3][AIS-3][BPC-3][RES-3]\n- drop non-existent PRD links; point to existing docs sections\n- correct Build & Test code block; remove stale metrics\n- add notes to consult CI artifacts for coverage/status

* docs(prd): replace legacy PRDs with concise PRD suite; align with self-node master-by-default and productionization plan [AIR-3][AIS-3][BPC-3][DOC-3]

* docs(prd): update PRD and legacy specs; add QA and roadmap refs @BIP-341 @BIP-174

* chore(scripts): add network auto-promotion and verification helpers @BIP-341 @BIP-174

* feat(hsm): stabilize HSM integration tests with timeouts/unlock; add dev simulator hooks @BIP-341

* test(core): gate network-bound tests and fix flakiness; align DAO/Bitcoin expectations @BIP-174 @BIP-341

* chore(build): update configs and editor settings; fix runtime reentry and governance API @BIP-341

* test(ml): minor test adjustment; add PRD keep file @BIP-341

* docs(prd): align with recent changes (HSM sim unlock/timeouts, env-honest QA, CLI aliases) @BIP-341 @BIP-174

* docs(readme): add CLI aliases, HSM simulator guidance, and environment-honest testing notes @BIP-341

* test(core): update DAO/Bitcoin/network/perf tests; environment-honest adjustments @BIP-341 @BIP-174

* test(env-honest): gate ML and system tests when CLIs/resources missing; avoid false failures

* test(env-honest): gate ML and system tests when CLIs/resources missing; avoid false failures

* docs(prd): align status/architecture/QA with env-honest test gating, branch rename, and resource checks

* docs(prd): align status/architecture/QA with env-honest test gating, branch rename, and resource checks

* chore(dev): switch docker-compose to testnet, add RPC shim (BitcoinRpcClient), env-honest RPC smoke test, and fix YAML indentation

* Update src/security/hsm/providers/software.rs


* Update src/layer2/production.rs

adding validation for credential strength or warning about insecure configurations in production.


* Update src/bitcoin/rpc/mod.rs

implementing a proper async wrapper or documenting the performance implications of this blocking approach.


* Update scripts/verify_implementation_status.sh

using more specific patterns


* Update src/layer2/production.rs

Placeholder cred or self for dev but still requires user tk set for mainnet


* chore(devops): fix compose YAML and default to testnet RPC; add overrides via ANYA_BITCOIN_RPC_URL

* feat(bitcoin): centralize RPC endpoint resolution + unify public defaults and env override precedence

* feat(bitcoin): centralize external endpoints (Electrum, explorer, Liquid) with env-driven resolver

* feat(bitcoin): centralize external endpoints (Electrum, explorer, Liquid) with env-driven resolver

* test(bitcoin): classify RPC smoke skips (400/401 unauth, conn refused) [BPC-3]

* test(bitcoin): classify RPC smoke skips (400/401 unauth, conn refused) [BPC-3]

* docs(security): record removal of cargo-geiger and RUSTSEC remediation; update dependency patch versions and environment-honest policy

* docs(ci): add frontmatter to PRDs, relax PR automation for integration branch, raise file change limit

* storage: autoconfig env doc + decentralized error variant cleanup; docs table fixes; ipfs minimal rebuild

* chore(quality-gates): enforce strict QA + add formatting, dependency drift, skip policy\n\nLabels: [AIR-3][AIS-3][BPC-3][RES-3][qa] \nVerification: cargo fmt --check PASS; clippy -D warnings PASS; unimplemented=0; todo=0; warnings=0\nEvidence: scripts/verify_implementation_status.sh --auto-run (partial)

* Update tests/test_utilities.rs


* Update tests/sectional_test_utils.rs


* Update src/storage/mod.rs


* Update tests/storage_router.rs


* Pending changes exported from your codespace

* Update tests/simplified_integration.rs


* feat(docs,deps): add C/C++ dependency audit, migration guide and quick fix scripts

Adds:
- docs/audit/cpp-dependencies-audit.md (automated high-level report)
- docs/audit/cpp-migration-guide.md (phased migration plan & benchmarks guidance)
- scripts/run_dependency_audit.sh (generates audit report)
- scripts/cpp-deps-quick-fix.sh (automated rustls + pure zstd update & build test)
- doc-manage 'audit' and 'index' commands for streamlined maintenance

Rationale:
Reduce native build surface (OpenSSL, zstd-sys, rocksdb, libgit2) and establish clear path to pure Rust alternatives (rustls, zstd crate, redb/sled, gix). Improves build speed, security posture, portability.

Next Steps:
1. Execute quick fix script & validate CI
2. Introduce feature flags for storage + git backend swaps
3. Benchmark pre/post build & runtime
4. Plan phased RocksDB and libgit2 migrations

Security: Prepares removal of frequent CVE sources (OpenSSL, libgit2) in favor of memory-safe Rust.


* feat(dev-env): add WSL F drive offload helper script with safe phased plan

* chore(devops): enhance F drive offload helper script with signing check and safer guidance

* chore(devops): add local secret scan helper to f_drive_offload script

Adds 'scan-secrets' command for heuristic private key block detection (no network).
Avoids committing auth material; instructs devs to use gitleaks for deep scan.


* ci(security): introduce secret scanning (gitleaks) + local workflow test + deep secret heuristics

Adds:
- .github/workflows/secret-scan.yml (gitleaks on push/PR)
- scripts/ci/test_workflows_local.sh (simulate core CI locally)
- Enhanced scripts/f_drive_offload.sh with scan-secrets-deep
- Pre-commit hook (.githooks/pre-commit) invoking scans & optional gitleaks protect

Usage:
 git config core.hooksPath .githooks
 scripts/ci/test_workflows_local.sh

Security: Prevents accidental secret leakage and enforces early detection.


* Update tests/dao/business_agents/Cargo.toml


* Update src/storage/mod.rs


* Update src/web/web5_adapter.rs


---------

+++ src/security/crypto/kdf.rs
@@ -130,7 +130,7 @@ pub fn scrypt(password: &[u8], params: &ScryptParams) -> Vec<u8> {
     // Convert scrypt params to equivalent PBKDF2 params
     let pbkdf2_params = Pbkdf2Params {
         salt: params.salt.clone(),
-        iterations: params.n as u32, // Use N as iterations
+        iterations: params.n, // Use N as iterations
         key_length: params.key_length,
     };
     pbkdf2(password, &pbkdf2_params)
